feat: upgrade linting rules
