+++ Makefile
@@ -3,6 +3,7 @@
 lint:
 	@echo "Linting Go code..."
 	@golangci-lint run
+	@echo "Linting complete!"
 
 # Set the default goal to lint (running 'make' will run 'make lint')
 .DEFAULT_GOAL := lint
