+++ binary-codec/serdes/field_id_codec.go
@@ -1,8 +1,6 @@
 package serdes
 
 import (
-	"encoding/hex"
-
 	"github.com/Peersyst/xrpl-go/binary-codec/definitions"
 )
 
@@ -28,27 +26,4 @@ func encodeFieldID(fieldName string) ([]byte, error) {
 		return append(b, 0, byte(fh.TypeCode), byte(fh.FieldCode)), nil
 	}
 	return nil, nil
-}
-
-// Returns the field name represented by the given field ID in hex string form.
-func decodeFieldID(h string) (string, error) {
-	b, err := hex.DecodeString(h)
-	if err != nil {
-		return "", err
-	}
-	if len(b) == 1 {
-		return definitions.Get().GetFieldNameByFieldHeader(definitions.CreateFieldHeader(int32(b[0]>>4), int32(b[0]&byte(15))))
-	}
-	if len(b) == 2 {
-		firstByteHighBits := b[0] >> 4
-		firstByteLowBits := b[0] & byte(15)
-		if firstByteHighBits == 0 {
-			return definitions.Get().GetFieldNameByFieldHeader(definitions.CreateFieldHeader(int32(b[1]), int32(firstByteLowBits)))
-		}
-		return definitions.Get().GetFieldNameByFieldHeader(definitions.CreateFieldHeader(int32(firstByteHighBits), int32(b[1])))
-	}
-	if len(b) == 3 {
-		return definitions.Get().GetFieldNameByFieldHeader(definitions.CreateFieldHeader(int32(b[1]), int32(b[2])))
-	}
-	return "", nil
 }
